+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 See [commit messages](https://github.com/salesking/sk_api_schema/commits/) for details.
 Also signup to our [Developer Newsletter](http://www.salesking.eu/dev/newsletter/) to stay up-to-date !!!
 
+##2013-03
+
+* add path method to schema: SK::Api::Schema.path
+
 ##2013-01
 
 Contacts/Client changes